Add empty go.sum file
Github actions build step for go complains if there is no go.sum file.
Go tooling doesn't create a go.sum file if the module has no
dependencies as is currently the case for pipesore.

To get around this an empty go.sum file was created and added to the
repo.